requirements.txt
Django>=4.0,<4.1
psycopg2-binary>=2.9
django-filter>=22.1
python-dotenv>=1.0
djangorestframework>=3.13
djangorestframework-simplejwt>=5.2
celery>=5.2
django-celery-results>=2.4
django-redis>=5.3
django-debug-toolbar>=4.1
django-cors-headers>=4.3
cryptography>=41.0
pandas>=1.5
requests>=2.28
